requirements.txt
alembic==1.18.4
blinker==1.9.0
click==8.3.1
colorama==0.4.6
Flask==3.1.3
greenlet==3.3.2
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
psycopg==3.3.3
psycopg-binary==3.3.3
python-dotenv==1.2.2
SQLAlchemy==2.0.48
typing_extensions==4.15.0
tzdata==2025.3
waitress==3.0.2
Werkzeug==3.1.6
